requirements.txt
dash==2.18.2
dash-bootstrap-components==1.6.0
dash-iconify==0.1.2
pandas>=2.2.3
numpy>=1.24
pyarrow>=15.0.0
databricks-sql-connector[pyarrow]>=4.0.0
databricks-sdk>=0.60.0
python-dotenv>=1.0.1